[Refactor] Replace the components with those defined in XRPrimer (#115)
1.  Import data structures `Keypoints`  `Limbs`  and their transfom functions from XRPrimer. Remove the definition in XRMoCap and leave a warning log.
2. Do not install or test mmtrack by default, because its requirements is too strict.
3. Use visualization in XRPrimer for `MviewMpersonDataVisualization`.
4. Fix the wrong frame number error in `PanopticDataCovnerter`
5. Remove `xrmocap/io/h5py_helper.py` which is not used anywhere.
6. Fix outdated numpy import (bool, int, float), but numpy<1.24 is still required by package smplx.
7. Update dockerfile and docker build script.

+++ xrmocap/transform/keypoints3d/optim/fourdag_optimization.py
@@ -623,7 +623,8 @@ class FourDAGOptimizer(FourDAGBaseOptimizer):
                              kps2d_target[:,
                                           view * self.limb_info.get_kps_number(
                                           ):self.limb_info.get_kps_number() *
-                                          (view + 1)][2].T > 0).astype(np.int))
+                                          (view + 1)][2].T > 0).astype(
+                                              np.int32))
                     for kps_id in range(self.limb_info.get_kps_number()):
                         if corr_cnt[kps_id] <= 1:
                             kps_confidence[kps_id] = 0
